add protocol for typing
